Make repeater gfx driver compile again - not tested
[git-p4: depot-paths = "//depot/qt/main/": change = 57481]

+++ src/kernel/qgfxvfb_qws.cpp
@@ -401,7 +401,7 @@ bool QVFbScreen::connect( const QString &displaySpec )
 	shmrgn = (unsigned char *)shmat( shmId, 0, 0 );
 
     if ( (int)shmrgn == -1 || shmrgn == 0 ) {
-	qDebug("No shmrgn %d",shmrgn);
+	qDebug("No shmrgn %d", (int)shmrgn);
 	return FALSE;
     }
